rename model directory as resources
