+++ backend/src/api/index.js
@@ -2,9 +2,9 @@ const { Router } = require('express');
 
 const router = new Router();
 
-router.use('/documents', require('./model/documents'));
-router.use('/categories', require('./model/categories'));
-router.use('/users', require('./model/users'));
+router.use('/documents', require('./resources/documents'));
+router.use('/categories', require('./resources/categories'));
+router.use('/users', require('./resources/users'));
 
 
 module.exports = router;
