+++ backend/src/api/resources/categories/index.js
